Rename data_extractors to areena_api
